update: add new structure to nb utils
