+++ src/py/mat3ra/notebooks_utils/__init__.py
@@ -1,0 +1,31 @@
+from .core.api.settings import (
+    ACCOUNT_ID,
+    AUTH_TOKEN,
+    ENDPOINT_ARGS,
+    HOST,
+    MATERIALS_PROJECT_API_KEY,
+    ORGANIZATION_ID,
+    PORT,
+    SECURE,
+    VERSION,
+    absolute_path_to_settings_json_file,
+    settings_json_config,
+)
+from .ipython.ui import display_JSON
+from .settings import UPLOADS_FOLDER
+
+__all__ = [
+    "display_JSON",
+    "absolute_path_to_settings_json_file",
+    "settings_json_config",
+    "ACCOUNT_ID",
+    "AUTH_TOKEN",
+    "MATERIALS_PROJECT_API_KEY",
+    "ORGANIZATION_ID",
+    "PORT",
+    "SECURE",
+    "VERSION",
+    "HOST",
+    "ENDPOINT_ARGS",
+    "UPLOADS_FOLDER",
+]
