+++ src/py/mat3ra/notebooks_utils/core/api/settings.json
