+++ src/py/mat3ra/notebooks_utils/core/entity/job/analysis.py
