+++ src/py/mat3ra/notebooks_utils/core/entity/property/job.py
