+++ src/py/mat3ra/notebooks_utils/core/entity/workflow/api.py
@@ -1,0 +1,39 @@
+from mat3ra.api_client import APIClient, BankWorkflowEndpoints
+from mat3ra.wode import Workflow
+
+
+def get_or_create_workflow(api_client: APIClient, workflow: Workflow, owner_id: str) -> dict:
+    """
+    Creates a workflow in the collection if none with the same hash exists under the given owner.
+
+    Args:
+        api_client (APIClient): API client instance carrying the authorization context.
+        workflow: mat3ra-wode Workflow object.
+        owner_id (str): Account ID under which to search and create.
+
+    Returns:
+        dict: The workflow dict (existing or newly created).
+    """
+    existing = api_client.workflows.list({"hash": workflow.hash, "owner._id": owner_id})
+    if existing:
+        print(f"♻️  Reusing already existing Workflow: {existing[0]['_id']}")
+        return existing[0]
+    created = api_client.workflows.create(workflow.to_dict_without_special_keys(), owner_id=owner_id)
+    print(f"✅ Workflow created: {created['_id']}")
+    return created
+
+
+def copy_bank_workflow_by_system_name(endpoint: BankWorkflowEndpoints, system_name: str, account_id: str) -> dict:
+    """
+    Copies a bank workflow with given system name into the account's workflows.
+
+    Args:
+        endpoint (BankWorkflowEndpoints): an instance of BankWorkflowEndpoints class
+        system_name (str): workflow system name.
+        account_id (str): ID of account to copy the bank workflow into.
+
+    Returns:
+        dict: new account's workflow
+    """
+    bank_workflow_id = endpoint.list({"systemName": system_name})[0]["_id"]
+    return endpoint.copy(bank_workflow_id, account_id)["_id"]
