+++ src/py/mat3ra/notebooks_utils/ipython/_collab.py
@@ -5,15 +5,14 @@ from urllib.parse import unquote, urlparse
 import requests
 
 
+# TODO: remove alongside the COLLAB environment in NBs
 def get_notebook_info() -> dict:
     """
     Get the information about a currently running notebook in Google Colab.
-    Args:
-        None
+
     Return:
         a dict with notebook info.
     """
-    # ip = socket.gethostbyname(socket.gethostname())  # 172.28.0.12
     ip = os.getenv("COLAB_JUPYTER_IP")
     response = requests.get(f"http://{ip}:9000/api/sessions").json()[0]
 
@@ -36,11 +35,5 @@ def print_notebook_path() -> None:
     """
     A proxy function used for a single string return when
     the corresponding entry-point script in 'setup.py' is called.
-
-    Args:
-        None
-    Return:
-        None
     """
-    # 'return get_notebook_info()["notebook_path"]' returns a non-zero exit code; using 'print' instead.
     print(get_notebook_info()["notebook_path"])
