+++ src/py/mat3ra/notebooks_utils/material.py
@@ -1,0 +1,8 @@
+from .core.entity.material.io import get_materials, load_material_from_folder, load_materials_from_folder, set_materials
+
+__all__ = [
+    "get_materials",
+    "set_materials",
+    "load_materials_from_folder",
+    "load_material_from_folder",
+]
