+++ src/py/mat3ra/notebooks_utils/primitive/enums.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class EnvironmentsEnum(Enum):
+    PYODIDE = "pyodide"
+    PYTHON = "python"
+
+
+class SeverityLevelEnum(Enum):
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
